+++ package.json
@@ -24,7 +24,7 @@
     "@glazed/datamodel": "^0.3.0",
     "@glazed/devtools": "^0.2.0",
     "@glazed/did-datastore": "^0.3.1",
-    "@prisma/client": "^5.17.0",
+    "@prisma/client": "^3.15.0",
     "axios": "^0.27.2",
     "bcryptjs": "^2.4.3",
     "cors": "^2.8.5",
@@ -67,7 +67,7 @@
     "jest": "^29.5.0",
     "jest-mock-extended": "^2.0.4",
     "nodemon": "^2.0.16",
-    "prisma": "^5.17.0",
+    "prisma": "^3.15.2",
     "renamer": "^4.0.0",
     "supertest": "^6.3.3",
     "swagger-jsdoc": "^6.2.8",
