+++ prisma/schema.prisma
@@ -94,7 +94,7 @@ model Claim {
   // splat NormalizedRating
   aspect String?
   score  Float? // -1 to 1
-  stars  Int? // 0 to 5
+  stars  Int?   // 0 to 5
 
   // splat Measure
   amt         Float?
@@ -127,65 +127,6 @@ model Claim {
   lastUpdatedAt DateTime @default(now())
 }
 
-// this model is to connect between the claim and the data that supports it
-// like the name of the claim, the image, the source, etc.
-// maybe to add another atts in future like the urls that can be related to the claim, like the website, the social media, etc.
-model  ClaimData {
-  id      Int @id @default(autoincrement())
-  claimId Int @unique
-
-  name  String
-  // image Image[]
-}
-
-// this model is to store the image of the claim
-model Image {
-  id      Int @id @default(autoincrement())
-  claimId Int  // @unique
-  // claimDataId Int
-
-  url             String
-  digestMultibase String?
-  metadata        Json?
-  effectiveDate   DateTime
-  createdDate     DateTime @default(now())
-  owner           String
-  signature       String
-
-  // claimData ClaimData @relation(fields: [claimDataId], references: [id])
-}
-
-
-// this is for validator service
-model CandidUserInfo {
-  id         Int    @id @default(autoincrement())
-  claimId    Int?   @unique
-  firstName  String
-  lastName   String
-  email      String
-  profileURL String
-}
-
-model ValidationRequest {
-  id         Int    @id @default(autoincrement())
-  createdAt DateTime @default(now())
-
-  context          String
-  validatorEmail   String
-  claimId          Int
-  validationStatus ValidationStatus @default(PENDING)
-  response         Boolean?         @default(false)
-  validationDate   DateTime?
-  statement        String?
-}
-
-enum ValidationStatus {
-  PENDING
-  COMPLETED
-  REJECTED
-  ABANDONED
-}
-
 enum IssuerIdType {
   DID
   ETH
@@ -206,4 +147,3 @@ enum HowKnown {
   OPINION
   OTHER
 }
-
