+++ src/controllers/index.ts
@@ -4,6 +4,10 @@ export {
   claimsFeed,
   claimGetById,
   claimSearch,
+  claimReport,
+  getNodeForLoggedInUser,
+  searchNodes,
+  getNodeById
 } from "./api.controller";
 export {
   login,
@@ -11,10 +15,3 @@ export {
   refreshToken,
   githubAuthenticator
 } from "./auth.controller";
-
-export {
-  claimReport,
-  getNodeForLoggedInUser,
-  searchNodes,
-  getNodeById
-} from "./node.controller";
